LayoutTests/http/tests/inspector/network/har/har-basic-expected.txt

Basic tests for HAR.


== Running test suite: HAR.Basic
-- Running test case: HAR.Basic.Empty
{
  "log": {
    "version": "1.2",
    "creator": {
      "name": "WebKit Web Inspector",
      "version": "<filtered>"
    },
    "pages": [
      {
        "startedDateTime": "",
        "id": "page_0",
        "title": "http://127.0.0.1:8000/inspector/network/har/har-basic.html",
        "pageTimings": {}
      }
    ],
    "entries": []
  }
}

-- Running test case: HAR.Basic.FakeResources
{
  "log": {
    "version": "1.2",
    "creator": {
      "name": "WebKit Web Inspector",
      "version": "<filtered>"
    },
    "pages": [
      {
        "startedDateTime": "",
        "id": "page_0",
        "title": "http://127.0.0.1:8000/inspector/network/har/har-basic.html",
        "pageTimings": {}
      }
    ],
    "entries": [
      {
        "pageref": "page_0",
        "startedDateTime": "2017-10-23T01:55:52.694Z",
        "time": 0,
        "request": {
          "method": "GET",
          "url": "https://example.com/fake.js",
          "httpVersion": "",
          "cookies": [],
          "headers": [
            {
              "name": "Test-Request-Header",
              "value": "Test Request Header Value"
            }
          ],
          "queryString": [],
          "headersSize": -1,
          "bodySize": -1
        },
        "response": {
          "status": 0,
          "statusText": "",
          "httpVersion": "",
          "cookies": [],
          "headers": [],
          "content": {
            "size": 0,
            "compression": 0,
            "mimeType": "text/javascript"
          },
          "redirectURL": "",
          "headersSize": -1,
          "bodySize": -1
        },
        "cache": {},
        "timings": {
          "blocked": -1,
          "dns": -1,
          "connect": -1,
          "ssl": -1,
          "send": 0,
          "wait": 0,
          "receive": 0
        }
      },
      {
        "pageref": "page_0",
        "startedDateTime": "2017-10-23T01:55:52.694Z",
        "time": 700,
        "request": {
          "method": "GET",
          "url": "https://example.com/fake.js",
          "httpVersion": "HTTP/1.1",
          "cookies": [],
          "headers": [
            {
              "name": "Test-Request-Header",
              "value": "Test Request Header Value"
            }
          ],
          "queryString": [],
          "headersSize": 100,
          "bodySize": 0
        },
        "response": {
          "status": 200,
          "statusText": "OK",
          "httpVersion": "HTTP/1.1",
          "cookies": [],
          "headers": [
            {
              "name": "Test-Response-Header",
              "value": "Test Response Header Value"
            }
          ],
          "content": {
            "size": 1234,
            "compression": 434,
            "mimeType": "text/javascript"
          },
          "redirectURL": "",
          "headersSize": 200,
          "bodySize": 800,
          "_transferSize": 1000
        },
        "cache": {},
        "timings": {
          "blocked": 100.00000000000009,
          "dns": 99.99999999999987,
          "connect": 99.99999999999987,
          "ssl": 49.99999999999982,
          "send": 100.00000000000009,
          "wait": 100.00000000000009,
          "receive": 99.99999999999987
        },
        "serverIPAddress": "12.34.56.78",
        "connection": "1",
        "_fetchType": "Network Load"
      }
    ]
  }
}

-- Running test case: HAR.Basic.Timing
PASS: HAR Timings for entry1 should add up to ~500ms before wait.
PASS: HAR Timings for entry2 should add up to ~500ms before wait.
PASS: HAR Timings for entry3 should add up to ~500ms before wait.
{
  "log": {
    "version": "1.2",
    "creator": "<filtered>",
    "pages": [
      {
        "startedDateTime": "",
        "id": "page_0",
        "title": "http://127.0.0.1:8000/inspector/network/har/har-basic.html",
        "pageTimings": {}
      }
    ],
    "entries": [
      {
        "pageref": "page_0",
        "startedDateTime": "2017-10-23T01:55:52.694Z",
        "time": 700,
        "request": "<filtered>",
        "response": "<filtered>",
        "cache": {},
        "timings": {
          "blocked": 500,
          "dns": -1,
          "connect": -1,
          "ssl": -1,
          "send": 0,
          "wait": 100.00000000000009,
          "receive": 99.99999999999987
        },
        "_fetchType": "Network Load"
      },
      {
        "pageref": "page_0",
        "startedDateTime": "2017-10-23T01:55:52.694Z",
        "time": 700,
        "request": "<filtered>",
        "response": "<filtered>",
        "cache": {},
        "timings": {
          "blocked": 100.00000000000009,
          "dns": 99.99999999999987,
          "connect": -1,
          "ssl": -1,
          "send": 300.00000000000006,
          "wait": 100.00000000000009,
          "receive": 99.99999999999987
        },
        "_fetchType": "Network Load"
      },
      {
        "pageref": "page_0",
        "startedDateTime": "2017-10-23T01:55:52.694Z",
        "time": 700,
        "request": "<filtered>",
        "response": "<filtered>",
        "cache": {},
        "timings": {
          "blocked": 300.00000000000006,
          "dns": -1,
          "connect": 99.99999999999987,
          "ssl": -1,
          "send": 100.00000000000009,
          "wait": 100.00000000000009,
          "receive": 99.99999999999987
        },
        "_fetchType": "Network Load"
      }
    ]
  }
}
